Fix: Strip trailing zeros from percentage and add month validation

+++ src/main/java/com/finance/manager/service/GoalService.java
@@ -118,7 +118,7 @@ public class GoalService {
 
         Map<String, Object> stats = new HashMap<>();
         stats.put("currentProgress", currentProgress);
-        stats.put("progressPercentage", percentage);
+        stats.put("progressPercentage", percentage.doubleValue());
         stats.put("remainingAmount", remainingAmount);
 
         return stats;
